Creation of service-keys (#842)
* Fetch state from aas-release-branch "service-key-creation"

Current top-commit-id of that branch: "9ca68503e2f08bd8123df39b5722decb0fcde1bb"

* Adapt gitignore and fix linting-message regarding java-indentation;

* 🔧 Fix `make clean`;

* 🔧 Fix certificate-location in tests;

* 🔧 Fix ginkgo-argument-order and use devbox-gingko!

* 🔧 Fix more test-cert-paths;

* 🔧 Fix output-style;

* 🔧 Fix error-matching-situation due to an erroneous implementation of the error-type;

* Make linter happy;

* Generate fakes only if not already present!

* 🔧 Fix tests;

* 🚧 Work in progress: Debug disk-usage-scenario;

Context: <https://github.com/cloudfoundry/app-autoscaler/actions/runs/19294343932/job/55191830755#step:6:1228>

* Add test-case for service-key-creation;

* Build gorouter-proxy if and only if necessary to speedup tests;

* Preparation for multiple schema-variants;

* Require schema-version when creating a service-key;

* Expect a schema-version in pure service-key-requests;

* Fix test-case;

* 🔧 Fix acceptance-tests

This commit does adaptions to a slightly modified output due to switching to a multi-version-schema.

* 🔧 Fix acceptance-tests;

* 🧹 Clean-up;

* Introduce test for misuse of the quite open legacy-schema to smuggle in a service-key-request;

* Enforce that old schema does not know the app_guid;

* Enforce legacy-schema to not offer a sensible ~schema-version~-Parameter;

Background: Distinguishability from the newer schemas which all should have that field filled out with sensible values.

* 🔧 Fix test;

* Re-ordering of test-cases;

* 🔧 Fix error in changed json-schema.

* 🛠️ Preparation for the client-communication within the broker;

* More concurrency-robustness for two tests;

* Add test-case for app-space-miss-alignment;

* 🔧 Fix new test for ensuring that the app is in the space.

* Introduce handling in case the cloud-controller does not provide binding-details;

* Make the linter happy;

* 🧹 Clean-up;

* Factor out helper-method;

* 🔧 Fixes two important issues

 + when the cloudcontroller provides an empty space-guid
 + when no app-guid is provided, the status-code 422 was sent which is reserved by the servicebrokerapi

* Rename schema-file according to proposal in PR-comment

See: <https://github.com/cloudfoundry/app-autoscaler/pull/842/files#r2560312068>

* Adds test for space-guid-comparison against the binding-db.

* 🔧 Fix test;

* Rename schema-file for version-id-inclusion;

* Fix potential nullpointer-issue;

* 🩹 Fix ~additionalProperties~-error;

* Switch to internal-server-error in case of a json-schema error-serialisation-issue;

* Improve readability of ~checkAppInSpace~;

* Allow only exactly one schema to be valid at a certain point in time.

* Improve comment and make validator-usage more understandable;

* Switch schema-version-numbering-start to "0.1" instead of "0.9"

See: <https://github.com/cloudfoundry/app-autoscaler/pull/842#discussion_r2567661697>

* Minor makefile-improvements;

* Add simple acceptance-test for service-key-creation;

* Make linter happy;

* Intentionally make the borker-acceptance-tests fail to see if they work!

* Add test-case to validate if application gets scaled as well via service-keys;

* Move scaling-test for service-keys to the app-tests;

* Fix syntactic error and minor changes;

* Fix import-error and shorten service-key-name;

* Fix error; instance-name overwritten;

* 🚧 Work in progress: Test with other space-name;

* Switch dynamic test to a disk-usage-test;

* Complete dynamic-scaling-tests for service-key scenarios and fix error in json-schema;

* 🔧 Fix error in configuration-parsing and experiment with acceptance-tests;

* 🔧 Fix acceptance-tests for the broker.

* Fix disk-usage-test;

* ⌛ Increase waiting-time;

* 🖍️ Fix typing errors;

+++ acceptance/broker/broker_test.go
@@ -63,36 +63,20 @@ func (s serviceInstance) name() string {
 
 var _ = Describe("AutoScaler Service Broker", func() {
 	var appName string
-
-	BeforeEach(func() {
-		appName = helpers.CreateTestApp(cfg, "broker-test", 1)
-	})
-
-	AfterEach(func() {
-		if os.Getenv("SKIP_TEARDOWN") == "true" {
-			fmt.Println("Skipping Teardown...")
-		} else {
-			Eventually(cf.Cf("app", appName, "--guid"), cfg.DefaultTimeoutDuration()).Should(Exit())
-			Eventually(cf.Cf("logs", appName, "--recent"), cfg.DefaultTimeoutDuration()).Should(Exit())
-			Expect(cf.Cf("delete", appName, "-f", "-r").Wait(cfg.CfPushTimeoutDuration())).To(Exit(0))
-		}
-	})
-
 	Context("performs lifecycle operations", func() {
-
 		var instance serviceInstance
 
 		BeforeEach(func() {
+			appName = helpers.CreateTestApp(cfg, "broker-test", 1)
+
 			instance = createService(cfg.ServicePlan)
 		})
-
 		It("fails to bind with invalid policies", func() {
 			bindService := cf.Cf("bind-service", appName, instance.name(), "-c", "../assets/file/policy/invalid.json").Wait(cfg.DefaultTimeoutDuration())
 			Expect(bindService).To(Exit(1))
 			combinedBuffer := gbytes.BufferWithBytes(append(bindService.Out.Contents(), bindService.Err.Contents()...))
-			Eventually(string(combinedBuffer.Contents())).Should(ContainSubstring(`[{"context":"(root).scaling_rules.1.adjustment","description":"Does not match pattern '^[-+][1-9]+[0-9]*%?$'"}]`))
+			Eventually(string(combinedBuffer.Contents())).Should(ContainSubstring(`{"context":"(root).scaling_rules.1.adjustment","description":"Does not match pattern '^[-+][1-9]+[0-9]*%?$'"}`))
 		})
-
 		It("binds&unbinds with policy", func() {
 			policyFile := "../assets/file/policy/all.json"
 			policy, err := os.ReadFile(policyFile)
@@ -106,7 +90,6 @@ var _ = Describe("AutoScaler Service Broker", func() {
 
 			instance.unbind(appName)
 		})
-
 		It("binds&unbinds with policy having credential-type as x509", func() {
 			policyFile := "../assets/file/policy/policy-with-credential-type.json"
 			_, err := os.ReadFile(policyFile)
@@ -150,16 +133,24 @@ var _ = Describe("AutoScaler Service Broker", func() {
 		})
 
 		AfterEach(func() {
-			instance.delete()
+			if os.Getenv("SKIP_TEARDOWN") == "true" {
+				fmt.Println("Skipping Teardown...")
+			} else {
+				Eventually(cf.Cf("app", appName, "--guid"), cfg.DefaultTimeoutDuration()).Should(Exit())
+				Eventually(cf.Cf("logs", appName, "--recent"), cfg.DefaultTimeoutDuration()).Should(Exit())
+				Expect(cf.Cf("delete", appName, "-f", "-r").Wait(cfg.CfPushTimeoutDuration())).To(Exit(0))
+
+				instance.delete()
+			}
 		})
 	})
-
 	Describe("allows setting default policies", func() {
 		var instance serviceInstance
 		var defaultPolicy []byte
 		var policy []byte
 
 		BeforeEach(func() {
+			appName = helpers.CreateTestApp(cfg, "broker-test", 1)
 			instance = createServiceWithParameters(cfg.ServicePlan, "../assets/file/policy/default_policy.json")
 			Expect(instance).NotTo(BeEmpty())
 			var err error
@@ -167,7 +158,7 @@ var _ = Describe("AutoScaler Service Broker", func() {
 			Expect(err).NotTo(HaveOccurred())
 
 			var serviceParameters = struct {
-				DefaultPolicy interface{} `json:"default_policy"`
+				DefaultPolicy any `json:"default_policy"`
 			}{}
 
 			err = json.Unmarshal(defaultPolicy, &serviceParameters)
@@ -197,13 +188,18 @@ var _ = Describe("AutoScaler Service Broker", func() {
 			if os.Getenv("SKIP_TEARDOWN") == "true" {
 				fmt.Println("Skipping Teardown...")
 			} else {
+				Eventually(cf.Cf("app", appName, "--guid"), cfg.DefaultTimeoutDuration()).Should(Exit())
+				Eventually(cf.Cf("logs", appName, "--recent"), cfg.DefaultTimeoutDuration()).Should(Exit())
+				Expect(cf.Cf("delete", appName, "-f", "-r").Wait(cfg.CfPushTimeoutDuration())).To(Exit(0))
 				instance.delete()
 			}
 		})
 	})
-
 	Describe("allows updating service plans", func() {
 		var instance serviceInstance
+		BeforeEach(func() {
+			appName = helpers.CreateTestApp(cfg, "broker-test", 1)
+		})
 		It("should update a service instance from one plan to another plan", func() {
 			servicePlans := GetServicePlans(cfg)
 			source, target, err := servicePlans.getSourceAndTargetForPlanUpdate()
@@ -211,9 +207,138 @@ var _ = Describe("AutoScaler Service Broker", func() {
 			instance = createService(source.Name)
 			instance.updatePlan(target.Name)
 		})
-
 		AfterEach(func() {
-			instance.delete()
+			if os.Getenv("SKIP_TEARDOWN") == "true" {
+				fmt.Println("Skipping Teardown...")
+			} else {
+				Eventually(cf.Cf("app", appName, "--guid"), cfg.DefaultTimeoutDuration()).Should(Exit())
+				Eventually(cf.Cf("logs", appName, "--recent"), cfg.DefaultTimeoutDuration()).Should(Exit())
+				Expect(cf.Cf("delete", appName, "-f", "-r").Wait(cfg.CfPushTimeoutDuration())).To(Exit(0))
+
+				instance.delete()
+			}
+		})
+	})
+	Context("Create a service-key", func() {
+		var serviceInstance serviceInstance
+		var serviceInstanceName string
+		When("providing a valid app-guid", func() {
+			When("the corresponding app isn't in the same space than the service-instance", func() {
+				var currentOrgName string
+				var currentSpaceName string
+				var otherSpaceName string
+				var appGuid string
+				BeforeEach(func() {
+					serviceInstance = createService(cfg.ServicePlan)
+					serviceInstanceName = string(serviceInstance)
+
+					currentSpaceName = setup.TestSpace.SpaceName()
+					otherSpaceName = currentSpaceName + "_other"
+					currentOrgName = setup.TestSpace.OrganizationName()
+					adminCtx := setup.AdminUserContext()
+					adminCtx.Org = currentOrgName
+					workflowhelpers.AsUser(adminCtx, cfg.DefaultTimeoutDuration(), func() {
+						cf.Cf("create-space", otherSpaceName, "-o", currentOrgName).Wait(cfg.DefaultTimeoutDuration())
+						cf.Cf("target", "-o", currentOrgName, "-s", otherSpaceName).Wait(cfg.DefaultTimeoutDuration())
+
+						appName = helpers.CreateTestApp(cfg, "broker-test", 1)
+						var err error
+						appGuid, err = helpers.GetAppGuid(cfg, appName)
+						Expect(err).ToNot(HaveOccurred())
+					})
+				})
+				AfterEach(func() {
+					if os.Getenv("SKIP_TEARDOWN") == "true" {
+						fmt.Println("Skipping Teardown...")
+					} else {
+						serviceInstance.delete()
+
+						adminCtx := setup.AdminUserContext()
+						adminCtx.Org = currentOrgName
+						adminCtx.Space = otherSpaceName
+						workflowhelpers.AsUser(adminCtx, cfg.DefaultTimeoutDuration(), func() {
+							s := cf.Cf("target", "-o", currentOrgName, "-s", otherSpaceName).Wait(cfg.DefaultTimeoutDuration())
+							Expect(s).To(Exit(0), "failed targeting org %s and space %s", currentOrgName, otherSpaceName)
+
+							Eventually(cf.Cf("app", appName, "--guid"), cfg.DefaultTimeoutDuration()).Should(Exit())
+							Eventually(cf.Cf("logs", appName, "--recent"), cfg.DefaultTimeoutDuration()).Should(Exit())
+							Expect(cf.Cf("delete", appName, "-f", "-r").Wait(cfg.CfPushTimeoutDuration())).To(Exit(0))
+
+							s = cf.Cf("delete-space", otherSpaceName, "-f", "-o", currentOrgName).Wait(cfg.DefaultTimeoutDuration())
+							Expect(s).To(Exit(0), "failed deleting space %s", otherSpaceName)
+						})
+					}
+				})
+
+				It("fails", func() {
+					// Preparation
+					paramsTemplate := `
+{
+  "schema-version": "0.1",
+  "configuration": {
+	"app_guid": "%s"
+  }
+}
+`
+					params := fmt.Sprintf(paramsTemplate, appGuid)
+
+					// Execution
+					serviceKeyName := fmt.Sprintf("%s@%s", appName, serviceInstanceName)
+					session := cf.Cf("create-service-key", serviceInstanceName, serviceKeyName, "-c", params).
+						Wait(cfg.DefaultTimeoutDuration())
+
+					// Validation
+					Expect(session).To(Exit(1))
+
+					combinedBuffer := gbytes.BufferWithBytes(append(session.Out.Contents(), session.Err.Contents()...))
+					expectedErrMsg := fmt.Sprintf("app %s not found in space", appGuid)
+					Eventually(string(combinedBuffer.Contents())).Should(ContainSubstring(expectedErrMsg))
+				})
+			})
+			When("the corresponding app is in the same space than the service-instance", func() {
+				var appGuid string
+				BeforeEach(func() {
+					appName = helpers.CreateTestApp(cfg, "broker-test", 1)
+					serviceInstance = createService(cfg.ServicePlan)
+					serviceInstanceName = string(serviceInstance)
+
+					var err error
+					appGuid, err = helpers.GetAppGuid(cfg, appName)
+					Expect(err).ToNot(HaveOccurred())
+				})
+				AfterEach(func() {
+					if os.Getenv("SKIP_TEARDOWN") == "true" {
+						fmt.Println("Skipping Teardown...")
+					} else {
+						Eventually(cf.Cf("app", appName, "--guid"), cfg.DefaultTimeoutDuration()).Should(Exit())
+						Eventually(cf.Cf("logs", appName, "--recent"), cfg.DefaultTimeoutDuration()).Should(Exit())
+						Expect(cf.Cf("delete", appName, "-f", "-r").Wait(cfg.CfPushTimeoutDuration())).To(Exit(0))
+
+						serviceInstance.delete()
+					}
+				})
+
+				It("succeeds on simple service-key creation", func() {
+					// Preparation
+					paramsTemplate := `
+{
+  "schema-version": "0.1",
+  "configuration": {
+	"app_guid": "%s"
+  }
+}
+`
+					params := fmt.Sprintf(paramsTemplate, appGuid)
+
+					// Execution
+					serviceKeyName := fmt.Sprintf("%s@%s", appName, serviceInstanceName)
+					session := cf.Cf("create-service-key", serviceInstanceName, serviceKeyName, "-c", params).
+						Wait(cfg.DefaultTimeoutDuration())
+
+					// Validation
+					Expect(session).To(Exit(0))
+				})
+			})
 		})
 	})
 })
